# Conflicts:
#	Cargo.toml

+++ crates/defi-abi/src/curve/pools.rs
@@ -115,7 +115,7 @@ sol! {
     #[derive(Debug, PartialEq, Eq)]
     interface ICurveU256_3_Eth{
         function get_dy(uint256,uint256,uint256) external view returns (uint256);
-        function calc_withdraw_one_coin(uint256,int128) external view returns (uint256);
+        function calc_withdraw_one_coin(uint256,uint256) external view returns (uint256);
         function calc_token_amount(uint256[3],bool) external view returns (uint256);
         function exchange(uint256,uint256,uint256,uint256,bool) external;
         function remove_liquidity_one_coin(uint256,uint256,uint256) external;
